backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic-settings==2.1.0
alembic==1.13.0
aiosmtplib==3.0.1
aioimaplib==1.1.0
openai==1.40.0
requests==2.31.0
APScheduler==3.10.4
email-validator==2.1.0
python-multipart==0.0.6
pymupdf==1.24.0
